// src/store/types/people.ts
export interface Person {
  name: string
  height: string
  mass: string
  hair_color: string
  skin_color: string
  eye_color: string
  birth_year: string
  gender: string
  homeworld: string
  films: string[]
  species: string[]
  vehicles: string[]
  starships: string[]
  created: string
  edited: string
  url: string
}

export interface PersonDisplayedInfo {
  name: string
  height: string
  mass: string
  hair_color: string
  skin_color: string
  eye_color: string
  birth_year: string
  gender: string
}

export interface PeopleState {
  nextPage: string | null
  previousPage: string | null
  people: PersonDisplayedInfo[]
  error: string | null
  isLoading: boolean
};